+++ packages/storefront/src/utils/generator/generator.tsx
@@ -45,6 +45,14 @@ import {
   type PInputEmailProps,
   PInputNumber,
   type PInputNumberProps,
+  PInputDate,
+  type PInputDateProps,
+  PInputMonth,
+  type PInputMonthProps,
+  PInputWeek,
+  type PInputWeekProps,
+  PInputTime,
+  type PInputTimeProps,
   PInputPassword,
   type PInputPasswordProps,
   PInputSearch,
@@ -257,6 +265,8 @@ export type PropTypeMapping = {
   'p-input-password': PInputPasswordProps;
   'p-input-number': PInputNumberProps;
   'p-input-date': PInputDateProps;
+  'p-input-month': PInputMonthProps;
+  'p-input-week': PInputWeekProps;
   'p-input-time': PInputTimeProps;
   'p-input-search': PInputSearchProps;
   'p-input-text': PInputTextProps;
@@ -331,6 +341,8 @@ const componentMap: Record<ConfiguratorTagNames, React.ElementType> = {
   'p-input-password': PInputPassword,
   'p-input-number': PInputNumber,
   'p-input-date': PInputDate,
+  'p-input-month': PInputMonth,
+  'p-input-week': PInputWeek,
   'p-input-time': PInputTime,
   'p-input-search': PInputSearch,
   'p-input-text': PInputText,